external_signer: replace boost::process with cpp-subprocess
This primarily affects the `RunCommandParseJSON` utility function.
Note that the arguments have to be passed now in form of a list instead
of a single string, i.e. the first parameter type changes from
`std::string` to `std::vector<std::string>` accordingly and all the
call-sides have to be adapted.